+++ .ps-rule/Rule.Rule.ps1
@@ -5,8 +5,8 @@
 Rule 'Rule.Name' -Type 'PSRule.Rules.Rule' {
     Recommend 'Rule name should be less than 35 characters to prevent being truncated.'
     Reason "The rule name is too long."
-    $TargetObject.RuleName.Length -le 35
-    $TargetObject.RuleName.StartsWith('CAF.')
+    $Assert.LessOrEqual($TargetObject, 'RuleName', 35);
+    $Assert.StartsWith($TargetObject, 'RuleName', 'CAF.');
 }
 
 # Synopsis: Complete help documentation
@@ -14,6 +14,19 @@ Rule 'Rule.Help' -Type 'PSRule.Rules.Rule' {
     $Assert.HasFieldValue($TargetObject, 'Info.Synopsis')
     $Assert.HasFieldValue($TargetObject, 'Info.Description')
     $Assert.HasFieldValue($TargetObject, 'Info.Recommendation')
+}
+
+# Synopsis: Use Microsoft Azure Well-Architected Framework pillars
+Rule 'Rule.AWAF' -Type 'PSRule.Rules.Rule' {
+    $Assert.HasFieldValue($TargetObject, 'Info.Annotations.category')
+    $Assert.HasFieldValue($TargetObject, 'Info.Annotations.pillar')
+    $Assert.In($TargetObject, 'Info.Annotations.pillar', @(
+        'Cost Optimization'
+        'Operational Excellence'
+        'Performance Efficiency'
+        'Reliability'
+        'Security'
+    ))
 }
 
 # Synopsis: Use online help